google-api/document_ai/pyproject.toml
[project]
name = "document-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "gcloud>=0.18.3",
    "google-cloud-documentai>=3.5.0",
    "python-dotenv>=1.1.1",
    "PyPDF2>=3.0.0",
    "reportlab>=4.0.0",
    "pillow>=10.0.0",
    "pdf2image>=1.17.0",
]
